backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-filter
python-dotenv
pandas
openpyxl
reportlab
